+++ applications/lokqlDx/ApplicationPreferencesWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="lokqlDx.PreferencesWindow"
+﻿<Window x:Class="lokqlDx.ApplicationPreferencesWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -6,7 +6,7 @@
         xmlns:local="clr-namespace:lokqlDx"
         mc:Ignorable="d"
         Title="Application Preferences" Height="450" Width="800"
-        Loaded="PreferencesWindow_OnLoaded">
+        Loaded="Dialog_OnLoaded">
   <DockPanel>
     <StackPanel Orientation="Horizontal" DockPanel.Dock="Bottom" HorizontalAlignment="Right">
       <Button Margin="3" IsCancel="True">Cancel</Button>
